requirements.txt
# Development
pyinstaller

# Dependencies
wxpython
numpy
scipy
opencv-python
pillow